refactor: move topology into solving
