+++ expertsystem/amplitude/canonical_decay.py
@@ -3,7 +3,7 @@
 from typing import Any, Callable, List, Optional
 
 from expertsystem.data import Spin
-from expertsystem.topology import StateTransitionGraph
+from expertsystem.solving.topology import StateTransitionGraph
 from expertsystem.types import InteractionProperties, ParticleWithSpin
 
 from .helicity_decay import (
