+++ expertsystem/solving/topology.py
