+++ expertsystem/state/combinatorics.py
@@ -20,7 +20,7 @@ from typing import (
 
 from expertsystem._utils import arange
 from expertsystem.data import Particle, ParticleCollection
-from expertsystem.topology import StateTransitionGraph, Topology
+from expertsystem.solving.topology import StateTransitionGraph, Topology
 from expertsystem.types import ParticleWithSpin
 
 StateWithSpins = Tuple[str, Sequence[float]]
