+++ tests/unit/state/test_combinatorics.py
@@ -6,6 +6,12 @@ from math import factorial
 import pytest
 
 from expertsystem import io
+from expertsystem.solving.topology import (
+    InteractionNode,
+    SimpleStateTransitionTopologyBuilder,
+    StateTransitionGraph,
+    Topology,
+)
 from expertsystem.state.combinatorics import (
     KinematicRepresentation,
     _generate_kinematic_permutations,
@@ -14,12 +20,6 @@ from expertsystem.state.combinatorics import (
     _safe_set_spin_projections,
     get_kinematic_representation,
     initialize_graph,
-)
-from expertsystem.topology import (
-    InteractionNode,
-    SimpleStateTransitionTopologyBuilder,
-    StateTransitionGraph,
-    Topology,
 )
